+++ config/app_config.py
@@ -5,8 +5,8 @@ Global application configuration
 
 import os
 
-VERSION = "2.6.4"
-UPDATE_DATE = "2026-07-03"
+VERSION = "2.6.5"
+UPDATE_DATE = "2026-08-18"
 APP_NAME = "PRISM"
 APP_FULL_NAME = "Plasma Research Integrated System for Multi-diagnostics"